refactor: fix blocking route by moving auth client-side and streaming featured threads
Move supabase.auth.getUser() from root layout to client-side AuthProvider
to prevent blocking the entire page in Next.js 16. Split featured threads
into a server component (data fetching) wrapped in Suspense inside a
client component (Embla carousel chrome).
