+++ app/src/app/(main)/page.tsx
@@ -1,19 +1,16 @@
-import { createClient } from "@/server/supabase/server";
-import { sessions } from "@/server/repos/sessions";
-
 import { HomePageHeroIntroduction } from "@/components/home-page-hero-introduction";
+import { HomePageFeaturedThreads } from "@/components/home-page-featured-threads";
 import { HomePageFeaturedThreadsCarousel } from "@/components/home-page-featured-threads-carousel";
 import { HomePageTutorialsList } from "@/components/home-page-tutorials-list";
 
-const HomePage = async () => {
-  const supabase = await createClient();
-  const threads = await sessions.getFeaturedThreads(supabase);
-
+const HomePage = () => {
   return (
     <div className="overflow-hidden [--util-grid-columns:10] md:[--util-grid-columns:20] lg:[--util-grid-columns:30]">
       <HomePageHeroIntroduction />
       <HomePageTutorialsList />
-      <HomePageFeaturedThreadsCarousel threads={threads} />
+      <HomePageFeaturedThreads>
+        <HomePageFeaturedThreadsCarousel />
+      </HomePageFeaturedThreads>
     </div>
   );
 };
